Create Macintosh.cpp
This Project uses graphics header file to create a mini  old macintosh with old school keyboard and mouse. Here the mouse of the macintosh can be moved using the user's mouse and this will move the word "MACINTOSH' in the screen of the macintosh. Few might face flicker due to low refresh rate, which can be adjusted as per wish. Right click or space will close the program.